ref(dashboard): Decompose dashboard components
Split dashboard UI into focused component files with Tailwind styling colocated at the rendering owner. Keep transcript grouping in a dashboard-local render model and preserve the existing dashboard API boundary.

Avoid dashboard-specific storage changes in Junior core while keeping turn chart tooltip detail in the dashboard client.
