+++ packages/junior-dashboard/src/client/components/JuniorLogo.tsx
@@ -1,0 +1,8 @@
+/** Render the compact Junior wordmark used by the dashboard shell. */
+export function JuniorLogo() {
+  return (
+    <div className="grid size-9 shrink-0 select-none place-items-center bg-black text-[0.82rem] font-black leading-none text-white">
+      Jr
+    </div>
+  );
+}
