+++ packages/junior-dashboard/src/tailwind.css
@@ -1,5 +1,13 @@
 @import "tailwindcss";
 
+@theme {
+  --font-sans:
+    -apple-system, "system-ui", "Segoe UI", Roboto, Oxygen, Ubuntu, sans-serif;
+  --font-mono:
+    "Geist Mono", ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas,
+    "Liberation Mono", monospace;
+}
+
 @source "./app.ts";
 @source "./client.tsx";
 @source "./client/**/*.{ts,tsx}";
